+++ examples/load.rs
@@ -7,7 +7,7 @@ fn main() {
     println!("Header: {:#?}", header);
     assert_head(header);
 
-    let levels = reader.levels().collect::<Vec<_>>();
+    let levels = reader.levels().map(|level| level.bytes).collect::<Vec<_>>();
     assert_eq!(levels.len(), header.level_count.max(1) as usize);
 
     let data = reader.data();